requirements.txt
numpy>=1.26.0
pyyaml>=6.0.2
rank-bm25>=0.2.2
duckdb>=1.1.3
fastembed>=0.7.4
faiss-cpu>=1.8.0
scikit-learn>=1.5.2
matplotlib>=3.9.0
tiktoken>=0.8.0
pypdf>=5.0.1
python-docx>=1.1.2
docling>=2.42.0
marker-pdf>=1.8.1
openai>=1.107.0
pytest>=8.3.3
notebook>=7.2.2
